+++ web-client/src/components/NotificationsDrawer/NotificationsDrawer.tsx
@@ -2,8 +2,8 @@ import { Drawer } from 'antd';
 import { firestore } from 'firebase';
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { setOffer } from 'src/ducks/specificOffers/actions';
-import { OffersState } from 'src/ducks/specificOffers/types';
+import { setOffer } from 'src/ducks/xxspecificOffers/actions';
+import { OffersState } from 'src/ducks/xxspecificOffers/types';
 import { Offer } from 'src/models/offers';
 import styled from 'styled-components';
