+++ web-client/src/ducks/ResponseOffers/reducer.ts
