+++ web-client/src/ducks/ResponseOffers/types.ts
