+++ web-client/src/ducks/xxspecificOffers/functions.ts
