+++ web-client/src/ducks/xxspecificOffers/reducer.ts
