+++ web-client/src/ducks/xxspecificOffers/types.ts
